+++ src/textual/widgets/_radio_set.py
@@ -37,6 +37,13 @@ class RadioSet(Container):
         This message can be handled using an `on_radio_set_changed` method.
         """
 
+        index: int
+        """The index of the radio button that was pressed to make the change."""
+        pressed: RadioButton
+        """The `RadioButton` that was pressed to make the change."""
+        sender: RadioSet
+        """The radio set that posted the message."""
+
         def __init__(self, sender: RadioSet, pressed: RadioButton) -> None:
             """Initialise the message.
 
@@ -45,17 +52,13 @@ class RadioSet(Container):
                 pressed: The radio button that was pressed.
             """
             super().__init__(sender)
-            self.input = sender
-            """A reference to the `RadioSet` that was changed."""
             self.pressed = pressed
-            """The `RadioButton` that was pressed to make the change."""
             # Note: it would be cleaner to use `sender.pressed_index` here,
             # but we can't be 100% sure all of the updates have happened at
             # this point, and so we can't go looking for the index of the
             # pressed button via the normal route. So here we go under the
             # hood.
             self.index = sender._nodes.index(pressed)
-            """The index of the `RadioButton` that was pressed to make the change."""
 
     def __init__(
         self,